requirements.txt
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
Flask==3.0.0
grpcio==1.59.0
grpcio-tools==1.59.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpickle==3.0.2
MarkupSafe==2.1.3
Pillow==10.1.0
protobuf==4.24.4
requests==2.31.0
urllib3==2.0.7
Werkzeug==3.0.0
